1. Finished baseline BMI filtering function.
2. Created a GDPR filtering function. There are patients who only have GDPR10, but have data from before the consent date.
From these patients, records are kept either if the consent date falls within the medical record start and end dates, or if the medical record starts after the consent date.

To be able to confidently perform this operation, genomics-related dates are added to medical records in the cleaning pipeline. It is noted that the gdpr10_date and the currently named 'genomics_test_date' (which could be renamed, as it is apparently the date the patient authorizes the use of their data in the system) are the same.

Currently, BMI filtering is done first, and GDPR filtering is done by dropping non-consented records and patients from the BMI-filtered database. The functions run in different cells independently now.

In a next iteration, these functions could be better integrated, to be executed together, or maybe even creating an intermediary database.

Creating a valid/invalid consent flag upstream in the cleaning pipeline is also strongly suggested, as it would make the filtering operations smoother.