Experimental creation of a copy of the paper1 pipeline in an effort to simplify and refactor the code.
Here, we subset the filtered db to only include measurements and medical records, from the patients that have GPDR4 and only focus on first records. This can be subsetted analysis-specifically.

A new I/O structure is proposed - pnk_db2_p1_in for the medical records, measurements, and the so-called 'survival analysis input', the time window-target weight loss analysis on which most analyses build. The analysis results will be saved to pnk_db2_p1_out.
The 'sa' table has not yet been created, as in the first intent, in hot weather, the code was running for over 50 minutes, and had to be stopped.