Changed the 'sa input table generation' script in the refactored paper1 pipeline trial, to make it faster and also to calculate time window datapoints closest to the time window center, not the baseline date, as done earlier.
This 'towards-baseline-tendency' is still present in the original notebook!

After, created a total comparison, with all wg causes and eb values. The script runs but it is not good - problem with numbers.

It is the end of the day, and I'm without brains to revise and correct the problems - future problems it is.