Created scripts for both between-cohort and within-cohort comparisons, making demographic, outcome, eating behavior and weight gain cause comparisons, where appropriate - between EB and total, WGC and EB cohorts, and within these cohorts, between demographic, eating behavior and weight gain cause groups.
The scripts' modularization has started, I've saved stats helper modules in an external file, importing them for each cell and specific analysis. This is a good precedent, identifying common code blocks, defining them as functions, and saving them externally in continuation will boost notebook readability, claritiy and maintainability. This is a future direction that must be kept in pipeline design.

The scripts at this point are not documented at all - which is important, but maybe enough once modularization is complete.

As a next step, building a semi-modular regression pipeline on the same token is a good direction to go, until the next supervisor meeting in 2 days, the 29th May.