Created a new RF predictor of 10% WL at 60d, but it performs terribly, much worse than the one created about a month or so ago.
Committing to save, but the performance issues need to be revised.

In this commit, also made some improvements to the WGC categorization script, that is at the moment of committing is still running. Added to extra rules to the prompt for stringency (regarding use of the 'none' category and prohibition of inferences), and tried using GPT4.1.

This too will be committed again after finishing the run and revising the output.