Succesfully created random forest predictions of 10% weight loss in 60 days on the subset of the patients that did reach that followup lenght (about half of the population).
However an important issue with the code needs to be corrected. For now, it defines success cases as those where days_to_10%_wl is under 60, but it is better to do this calculation as wl_60d_% > 10. The days_to_10%_wl column is too many times empty, so these 0 values may be identified as success cases - I'm not sure! This needs to be revised well, as this data might be totally wrong now.