Identified a CRITICAL ISSUE in logistic regressions: the outcome was looked at only WITHIN THE DEFINED TIME WINDOW (wl_Nd_% column), not LEADING UP TO the defined cutoff day. However, this means that if someone loses 10% at 30 days and stops the program, at 60 days (which is a range of 50-70), they will appear as failure cases.
The updated logistic regression - and similarly, random forest - modules will instead take the days_to_X%_wl column as an input, and if that contains a value up to the defined time window end + 5 days (for permissivity), the case is a success.

The currently updated logistic regression module is ran from the notebook, as for some reason, updating the functions in the regressions module does not seem to affect the eventual output. No time to troubleshoot all day today, with presentation deadline tomorrow.

This was a major issue, good spot on it. In the future, things like this need to be avoided.