Updated Random Forest classifiers.
Using days_to_10%_wl <= 65 to catch cases where the outcome was reached before 60+-10 days (as wl_60d_% uses that window, not capturing previous instances); as well as edge cases reaching the target a little after 60 days.

Saved the output to paper1_directory/rf_roc_fi.

Deleted the section with default settings, as it performs worse than the tuned model, not adding much to the results.