Filtered DB for Italian PNK CONNECT patients, through the pnk_data_wrangling notebook.
In patients table, IDs belonging to patients from Italy and with valid GDPR4 consent (irrespective of GDPR10) are identified.
Comorbidities, measurements and prescriptions are filtered by these IDs.
As medical records have only record IDs, that table is filtered by the record IDs retained in the prescriptions DB post filtering.
Alleles are not included in the filtering as the scope is the CONNECT study, not including CONNECT GEN.
The raw, spanish language, unstructured DB is filtered, with the different ID name occurrences all accounted for.
After filtering in SQL, db is written to Excel.