Added a Mann-Whitney U-test to descriptive_comparisons,
after realizing that non-parametric tests might be needed for comparing continuous data, as most of the variables show a non-normal distribution on violin plots when inspected visually.

Applied Mann-Whitney U-test (aka Wilcoxon signed rank test) to cluster comparisons as a 'trial' and found double the significant differences in comparison with Welch's t.